完善后台功能

+++ src/main/java/com/hhd/service/impl/OssServiceImpl.java
@@ -1,38 +1,44 @@
 package com.hhd.service.impl;
 
 import cn.hutool.core.date.DateTime;
-import com.aliyun.oss.*;
+import com.aliyun.oss.ClientException;
+import com.aliyun.oss.OSS;
+import com.aliyun.oss.OSSClientBuilder;
+import com.aliyun.oss.OSSException;
 import com.aliyun.oss.model.*;
 import com.aliyun.vod.upload.impl.UploadVideoImpl;
 import com.aliyun.vod.upload.req.UploadStreamRequest;
+import com.aliyun.vod.upload.req.UploadVideoRequest;
 import com.aliyun.vod.upload.resp.UploadStreamResponse;
+import com.aliyun.vod.upload.resp.UploadVideoResponse;
 import com.aliyuncs.DefaultAcsClient;
 import com.aliyuncs.vod.model.v20170321.DeleteVideoRequest;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.hhd.exceptionhandler.CloudException;
-import com.hhd.mapper.FileMapper;
 import com.hhd.pojo.domain.UCenter;
 import com.hhd.pojo.entity.Files;
 import com.hhd.service.IFileService;
 import com.hhd.service.IOssService;
 import com.hhd.service.IUCenterService;
-import com.hhd.utils.InitVodClient;
-import com.hhd.utils.MD5;
 import com.hhd.utils.InitOssClient;
+import com.hhd.utils.MD5;
 import com.hhd.utils.R;
+import org.apache.tomcat.util.http.fileupload.disk.DiskFileItem;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
-import wiremock.com.github.jknack.handlebars.Lambda;
+import org.springframework.web.multipart.commons.CommonsMultipartFile;
 
+import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.TimeUnit;
 
 import static com.hhd.utils.InitVodClient.initVodClient;
 
@@ -48,17 +54,15 @@ public class OssServiceImpl implements IOssService {
     @Autowired
     private IUCenterService uService;
 
-    @Async
     @Override
     public Map<String, Files> upload(MultipartFile file, Files files) {
         String objectName = new DateTime().toDateStr() + "/" + file.getOriginalFilename();
         files.setType(file.getOriginalFilename().substring(file.getOriginalFilename().lastIndexOf(".")).substring(1));
         files.setFileName(file.getOriginalFilename().substring(0, file.getOriginalFilename().indexOf(".")));
         files.setMd5(md5.getFileMd5String(file));
-
         String url = InitOssClient.HTTPS_PREFIX + InitOssClient.BUCKET_NAME + "." + InitOssClient.END_POINT + "/" + objectName;
         files.setUrl(url);
-        System.out.println(files);
+
         OSS ossClient = new OSSClientBuilder().build(InitOssClient.END_POINT, InitOssClient.ACCESS_KEY_ID, InitOssClient.ACCESS_KEY_SECRET);
         try {
             // 创建InitiateMultipartUploadRequest对象。
@@ -73,7 +77,7 @@ public class OssServiceImpl implements IOssService {
             // 每个分片的大小，用于计算文件有多少个分片。单位为字节。
             final long partSize = 1024 * 1024L;
 
-//            计算文件分片
+            // 计算文件分片数
             long fileLength = file.getSize();
             int partCount = (int) (fileLength / partSize);
             if (fileLength % partSize != 0) {
@@ -128,53 +132,67 @@ public class OssServiceImpl implements IOssService {
         return map;
     }
 
-    @Async
+
     @Override
     public Files uploadVideo(MultipartFile file, Files files) {
-        String fileName = file.getOriginalFilename();
-        String title = fileName.substring(0, fileName.lastIndexOf("."));
-        files.setType(fileName.substring(fileName.lastIndexOf(".")).substring(1));
-        files.setFileName(title);
-        files.setMd5(md5.getFileMd5String(file));
-
+        LambdaQueryWrapper<UCenter> lqw = new LambdaQueryWrapper<>();
+        UCenter one = uService.getOne(lqw.eq(UCenter::getId, files.getUserId()));
+        Path pathCreate;
         try {
-            InputStream inputStream = file.getInputStream();
-            UploadStreamRequest request = new UploadStreamRequest(InitOssClient.ACCESS_KEY_ID,
-                    InitOssClient.ACCESS_KEY_SECRET, title, fileName, inputStream);
-
-            UploadVideoImpl uploader = new UploadVideoImpl();
-            UploadStreamResponse response = uploader.uploadStream(request);
-
-            String videoId = null;
-            if (response.isSuccess()) {
-                videoId = response.getVideoId();
-                System.out.println(response.getRequestId());
-            } else { //如果设置回调URL无效，不影响视频上传，可以返回VideoId同时会返回错误码。
-                // 其他情况上传失败时，VideoId为空，此时需要根据返回错误码分析具体错误原因
-                videoId = response.getVideoId();
-            }
-            files.setVideoId(videoId);
-            return files;
+            pathCreate = java.nio.file.Files.createDirectories(Paths.get(one.getDownLoadAdd() + "\\temp"));
         } catch (Exception e) {
-            e.printStackTrace();
-            return null;
+            throw new CloudException(R.ERROR, R.EXECUTION_ERR);
         }
+        files.setMd5(md5.getFileMd5String(file));
+        String name = file.getOriginalFilename();
+        String title = name.substring(0, name.lastIndexOf("."));
+        files.setType(name.substring(name.lastIndexOf(".")).substring(1));
+        files.setFileName(title);
+        String fileName = pathCreate.toString() + file.getOriginalFilename();
+        //上传文件名称需为上传文件绝对路径，MultipartFile无法直接获取，转储指定地址
+        File f = new File(fileName);
+        try {
+            file.transferTo(f);
+        } catch (Exception e) {
+            throw new CloudException(R.ERROR, R.EXECUTION_ERR);
+        }
+        UploadVideoRequest request = new UploadVideoRequest(InitOssClient.ACCESS_KEY_ID, InitOssClient.ACCESS_KEY_SECRET, title, fileName);
+        request.setPartSize(2 * 1024 * 1024L);
+        request.setTaskNum(5);
+        request.setEnableCheckpoint(true);
+        UploadVideoImpl uploader = new UploadVideoImpl();
+        UploadVideoResponse response = uploader.uploadVideo(request);
+        if (response.isSuccess()) {
+            files.setVideoId(response.getVideoId());
+            System.out.print("VideoId=" + response.getVideoId() + "\n");
+        } else {
+            /* 如果设置回调URL无效，不影响视频上传，可以返回VideoId同时会返回错误码。其他情况上传失败时，VideoId为空，此时需要根据返回错误码分析具体错误原因 */
+            System.out.print("VideoId=" + response.getVideoId() + "\n");
+            System.out.print("ErrorCode=" + response.getCode() + "\n");
+            System.out.print("ErrorMessage=" + response.getMessage() + "\n");
+        }
+        f.delete();
+        return files;
     }
 
 
     @Override
     public R delete(String id) {
         LambdaQueryWrapper<Files> lqw = new LambdaQueryWrapper<>();
-        lqw.eq(Files::getId, id);
-        Files one = fService.getOne(lqw);
-        String time = one.getCreateTime();
+        Files exist = fService.getOne(lqw.eq(Files::getId, id));
+        String time = exist.getCreateTime();
         String createTime = new DateTime(time).toDateStr();
-        String fileName = one.getFileName();
+        String fileName = exist.getFileName();
+        String type = exist.getType();
         fService.delById(id);
+        LambdaQueryWrapper<Files> lqw1 = new LambdaQueryWrapper<>();
+        if (fService.count(lqw1.eq(Files::getMd5, exist.getMd5())) != 0L) {
+            return R.ok();
+        }
         OSS ossClient = new OSSClientBuilder().build(InitOssClient.END_POINT,
                 InitOssClient.ACCESS_KEY_ID, InitOssClient.ACCESS_KEY_SECRET);
         try {
-            String fileKey = createTime + "/" + fileName;
+            String fileKey = createTime.substring(0, 10) + "/" + fileName + "." + type;
             ossClient.deleteObject(InitOssClient.BUCKET_NAME, fileKey);
         } catch (Exception e) {
             return R.error();
@@ -190,6 +208,10 @@ public class OssServiceImpl implements IOssService {
         Files one = fService.getOne(lqw.eq(Files::getId, id));
         String videoId = one.getVideoId();
         fService.delById(id);
+        LambdaQueryWrapper<Files> lqw1 = new LambdaQueryWrapper<>();
+        if (fService.count(lqw1.eq(Files::getMd5, one.getMd5())) != 0L) {
+            return R.ok();
+        }
         try {
             //初始化对象
             DefaultAcsClient client = initVodClient();
@@ -217,8 +239,12 @@ public class OssServiceImpl implements IOssService {
         UCenter user = uService.getOne(lqw1.eq(UCenter::getId, one.getUserId()));
         String objectName = one.getUrl().substring(47);
 
-        System.out.println(objectName);
-
+        //创建用户自定义的下载目录文件夹
+        try {
+            java.nio.file.Files.createDirectories(Paths.get(user.getDownLoadAdd()));
+        } catch (Exception e) {
+            throw new CloudException(R.ERROR, R.EXECUTION_ERR);
+        }
         try {
             DownloadFileRequest downloadFileRequest = new DownloadFileRequest(InitOssClient.BUCKET_NAME, objectName);
             // 指定Object下载到本地文件的完整路径
